requirements.txt
numpy~=1.23.4
torchmetrics~=0.11.0
torchvision~=0.14.0
timm~=0.6.12
wandb~=0.13.6